requirements.txt
python-telegram-bot==20.7
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9
python-dotenv==1.0.1
reportlab==4.1.0
fastapi==0.110.2
uvicorn==0.29.0
openpyxl==3.1.2
python-jose[cryptography]
python-multipart
jinja2==3.1.3
